some eslint issues
